Refactor NFT modal feature names
